Replace ack plugin with ripgrep, tidy up layout
I haven't really use the ack plugin much so remove that. Replace it with
a similar plugin specific for ripgrep which is a thin wrapper around rg.

The layout of the plugins was very tight and hard to navigate. This
change spaces them out.

The order the plugins are loaded is grouped by similar function to make
them a bit more logical. i.e. all syntax related plugins are together.
And groups that affect more of vim or are more lower level go near the top.

The settings are re-arranged to match the plugin load order to make them
eaiser to navigate.

Added some quick reference text to some of the plugins so I can refresh
myself how to use them and they don't go unused.